+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Added endpoint to resolve ENS names to addresses (`GET /ens-address`)
+
 ## [0.4.0] - 2026-02-02
 
 ### Changed